fix: correct calculation of final Matrices in recursive algorithms
In this fix, I corrected the divAndConqMult() and strassensMult() algorithms: In
both, (2dArray).clone() statements were replaced with private helper method copyOf()
statements. Additionally, there was an error in the calculation of the "U" multiplication
in strassens algorithm that was fixed. Finally, in both algorithms, I corrected the final
calculation statement, that adds all the "c" sub-matrices, to work correctly and not
throw an IndexOutOfBoundsException.

Also, included a main for testing purposes.